Fix review points
