+++ waku/test_data.go
@@ -8,13 +8,6 @@ var DefaultWakuConfig WakuConfig
 
 func init() {
 
-	udpPort, _, err1 := GetFreePortIfNeeded(0, 0)
-	tcpPort, _, err2 := GetFreePortIfNeeded(0, 0)
-
-	if err1 != nil || err2 != nil {
-		Error("Failed to get free ports %v %v", err1, err2)
-	}
-
 	DefaultWakuConfig = WakuConfig{
 		Relay:           false,
 		LogLevel:        "DEBUG",
@@ -25,12 +18,13 @@ func init() {
 		Store:           false,
 		Filter:          false,
 		Lightpush:       false,
-		Discv5UdpPort:   udpPort,
-		TcpPort:         tcpPort,
+		Discv5UdpPort:   0,
+		TcpPort:         0,
 	}
 }
 
 const ConnectPeerTimeout = 10 * time.Second //default timeout for node to connect to another node
+const DefaultTimeOut = 3 * time.Second
 
 var DefaultPubsubTopic = "/waku/2/rs/16/64"
 var (
